remove a few circular refs
