+++ src/lib/data/miscRequirements.ts
@@ -1,0 +1,12 @@
+import { Skills } from '../types';
+
+export const soteSkillRequirements: Skills = {
+	mining: 70,
+	smithing: 70,
+	farming: 70,
+	woodcutting: 70,
+	agility: 70,
+	herblore: 70,
+	construction: 70,
+	hunter: 70
+};
