+++ src/lib/tickers.ts
@@ -8,7 +8,7 @@ import { mahojiUserSettingsUpdate } from '../mahoji/settingsUpdate';
 import { BitField, Channel, informationalButtons } from './constants';
 import { collectMetrics } from './metrics';
 import { prisma, queryCountStore } from './settings/prisma';
-import { runCommand } from './settings/settings';
+import { runCommand } from './settings/runCommand';
 import { getFarmingInfo } from './skilling/functions/getFarmingInfo';
 import Farming from './skilling/skills/farming';
 import { completeActivity } from './Task';
