+++ src/lib/util.ts
@@ -36,7 +36,6 @@ import { production, SupportServer } from '../config';
 import { skillEmoji, usernameCache } from './constants';
 import { DefenceGearStat, GearSetupType, GearSetupTypes, GearStat, OffenceGearStat } from './gear/types';
 import { Consumable } from './minions/types';
-import { MUserClass } from './MUser';
 import { PaginatedMessage } from './PaginatedMessage';
 import { POHBoosts } from './poh';
 import { SkillsEnum } from './skilling/types';
@@ -609,11 +608,7 @@ export async function asyncGzip(buffer: Buffer) {
 	});
 }
 
-export function skillingPetDropRate(
-	user: MUserClass,
-	skill: SkillsEnum,
-	baseDropRate: number
-): { petDropRate: number } {
+export function skillingPetDropRate(user: MUser, skill: SkillsEnum, baseDropRate: number): { petDropRate: number } {
 	const twoHundredMillXP = user.skillsAsXP[skill] >= 200_000_000;
 	const skillLevel = user.skillsAsLevels[skill];
 	const petRateDivisor = twoHundredMillXP ? 15 : 1;
